Fixed jsw_hresize(), which didn't work at all.
jsw_hresize() would build a new table, free the old table, and then
not do anything with the new table, so the caller was left holding an
invalid pointer to the old table.

jsw_hresize() ended with:
  htab = new_htab;
but since the htab variable went out of scope as soon as the function
returned, the new htab was leaked.

Also, even if it had worked, it was inefficient, because it copied all
the keys and items, which is unnecessary.

My new version of jsw_hresize() is correct, and also reuses as much of
the old table as possible, to avoid unnecessary allocation and
copying.

Here is the valgrind failure of the old version:

valgrind -q --error-exitcode=99 --exit-on-first-error=yes ./test-hlib 2963426107
test-hlib: seed = 2963426107
==542609== Invalid read of size 8
==542609==    at 0x109560: jsw_hfind (jsw_hlib.c:149)
==542609==    by 0x10A099: lookup_item (test-hlib.c:73)
==542609==    by 0x10A14F: step (test-main.c:37)
==542609==    by 0x10A384: main (test-main.c:89)
==542609==  Address 0x4a594a8 is 40 bytes inside a block of size 88 free'd
==542609==    at 0x483CA3F: free (in /usr/lib/x86_64-linux-gnu/valgrind/vgpreload_memcheck-amd64-linux.so)
==542609==    by 0x109544: jsw_hdelete (jsw_hlib.c:139)
==542609==    by 0x109B24: jsw_hresize (jsw_hlib.c:301)
==542609==    by 0x10A0C2: resize_container (test-hlib.c:78)
==542609==    by 0x10A399: main (test-main.c:91)
==542609==  Block was alloc'd at
==542609==    at 0x483B7F3: malloc (in /usr/lib/x86_64-linux-gnu/valgrind/vgpreload_memcheck-amd64-linux.so)
==542609==    by 0x109348: jsw_hnew (jsw_hlib.c:82)
==542609==    by 0x109FEA: new_container (test-hlib.c:51)
==542609==    by 0x10A336: main (test-main.c:81)
==542609==
==542609==
==542609== Exit program on first error (--exit-on-first-error=yes)